Fixed a few auth/validation issues

+++ src/main/java/com/magtable/controller/UserController.java
@@ -14,6 +14,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -25,7 +26,7 @@ public class UserController {
     @Autowired
     private UserRepository userRepository;
 
-   @Autowired
+    @Autowired
     private RoleRepository roleRepository;
 
     @Autowired
@@ -53,7 +54,7 @@ public class UserController {
     }
 
     /**
-     * route            GET /user/{id}
+     * route           GET /user/{id}
      * description     Get user by ID
      * access          Private - System Managers
      *
@@ -103,7 +104,7 @@ public class UserController {
             return user;
         } catch (DataIntegrityViolationException e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Username already exists");
-        }catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -118,19 +119,29 @@ public class UserController {
      * @param userId id of user to delete
      */
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable(value = "id") final int userId) {
-        // TODO make sure user isn't deleting themselves
+    public ResponseEntity<?> deleteUser(HttpServletRequest request, @PathVariable(value = "id") final int userId) {
         // need access to request username from JWT
+        String jwt = request.getHeader("Authorization");
         try {
+            //extracting the username from the jwt token
+            String username = jwtTokenUtil.extractUsername(jwt.substring(7)); // jwt can potentially be null
+            //searching the database for the user
+            User user = userRepository.findUserByUsername(username).orElseThrow(() ->
+                    new ResponseStatusException(HttpStatus.NOT_FOUND, String.format("User %s not found.", username)));
+
+            if (user.getId() == userId) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Cannot Delete Yourself.");
+            }
+
             userRepository.deleteById(userId);
             return ResponseEntity.ok(HttpStatus.OK);
         } catch (EmptyResultDataAccessException e) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND,
                     String.format("Deletion failed: User #%d not found.", userId));
+        } catch (NullPointerException e) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Authentication failed: JWT Not Found.");
         }
     }
-
-
 
 
     /**
@@ -155,7 +166,22 @@ public class UserController {
      * @return User object with newly reset password
      */
     @PutMapping("/reset/{id}")
-    public User resetPassword(@PathVariable(value = "id") final int userId) {
+    public User resetPassword(HttpServletRequest request, @PathVariable(value = "id") final int userId) {
+        try {
+            String jwt = request.getHeader("Authorization");
+            String username = jwtTokenUtil.extractUsername(jwt.substring(7)); // jwt can potentially be null todo address duplicate code error
+            User requestUser = userRepository.findUserByUsername(username).orElseThrow(() ->
+                    new ResponseStatusException(HttpStatus.NOT_FOUND, String.format("User %s not found.", username)));
+            // don't allow the requesting user to reset their own password in this way
+            System.out.println(requestUser.getId() + " " + userId);
+            if (requestUser.getId() == userId) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Cannot Reset Your Own Password.");
+            }
+        } catch (NullPointerException e) {
+            e.printStackTrace(); // TODO find out what will be thrown, update in future pls
+        }
+
+
         User user = userRepository.findById(userId).orElseThrow(() ->
                 new ResponseStatusException(HttpStatus.NOT_FOUND, String.format("User #%d not found.", userId)));
